requirements.txt
langchain==1.2.10
langchain-classic
langchain-community
langchain-core
langchain-groq
langchain-chroma
langchain-huggingface
pypdf
streamlit==1.2.10
python-dotenv